+++ network/fitness_calculator_cpu.cuh
@@ -6,8 +6,14 @@
 #define GA_CNN_FITNESS_CALCULATOR_CPU_CUH
 
 
-void calculateFitnessCPU(const int *labels, const float *images, const float *networks, int networkCount, int dataCount,
-                         int *fitness);
+void calculateFitnessCPU(
+        const int *labels,
+        const float *images,
+        const float *networks,
+        int networkCount,
+        int dataCount,
+        float *fitness
+);
 
 
 #endif //GA_CNN_FITNESS_CALCULATOR_CPU_CUH
