+++ network/fitness_calculator_gpu.cuh
@@ -7,6 +7,6 @@
 
 
 void calculateFitnessGPU(const int *labels, const float *images, const float *networks, int networkCount, int dataCount,
-                         int *fitness);
+                         float *fitness);
 
 #endif //GA_CNN_FITNESS_CALCULATOR_GPU_CUH
